Supplement the introduction
Add 
" HEXO does not need to be deployed to our server. The information stored on our server is actually based on the article written in HEXO through markdown, then HEXO helps us generate a static html page. Finally, it uploads the generated html to Our server is simple: HEXO is a static page generation, uploading tool."